SCRD-1759 Stop registering keystone v2.0 endpoint
The keystone v2 API is deprecated and it is being removed in Queens.
Using it causes deprecation warnings to be emitted in the keystone logs.
It was also causing warnings in the horizon logs:

  WARNING openstack_auth.backend The Keystone URL in service catalog
  points to a v2.0 Keystone endpoint, but v3 is specified as the API
  version to use by Horizon. Using v3 endpoint for authentication.

While there is a lot more work to do to remove v2 usage in the deployer
and in the other services, this patch at least helps clean up the
horizon logs and encourages end users to use the new endpoint. Note that
removing the endpoint from the catalog does not prevent it from being
used, so this will not break anything still relying on it.

Change-Id: Ic6b2619c938dc7e1bf5f8e0f08fa9a4910ddf891