SCRD-2641 Publish descriptions of input models
Update the descriptions in the README.md of each input model.
These descriptions are displayed to the user in the installer GUI.
Remove .html files since they are difficult to keep in sync with the .md
files, and should be generated programatically from the .md sources.
Comment out the copyright in each README to be consistent with other
source code and to avoid making the rendered HTML hard to read.
Render NIC names and filenames as fixed width.

Change-Id: Ia6dc0e3967829c2a75b829829cbe726ddb4b0b90