SCRD-2542 Modify input model for esx based models
The examples for Entry-scale-kvm-esx-ses and
Entry-scale-kvm-esx-ses-mml contain refereneces to servers,
and variables which will not be defined until the user takes
manual actions.
It is best to comment out those sections of the input model
and let the user ucomment them once the necessary manual changes
have been made to the system.
If the sections are NOT uncommented, and the user runs site.yml
using the input model 'as is' from example folder, the run will fail.

Change-Id: I865d824056ed5f1116e82a7ef9b065eee4a933c9