lock out /server-status calls to management network only (bsc#1068734)
/server-status on the ext network for some services provide more data
than needed and it was feared that it could be used by attackers.

Many of the other services result in 4XX http statuses.  A few provide
some API info.  barbican, horizon, and opsconsole provide apache status
information and this patch locks /server-status calls to these services
to the management network access only.

Change-Id: I12bc14c61e5a7952563a7242a52e3009f6266b88