SCRD-2891 deployer-as-controller input model with 3 CP nodes
A deployer-in-cloud input model with 3 CP nodes is needed
to test the maintenance update workflow scenario that
has the deployer colocated with one of the active cloud
nodes. 3 controller nodes are required to support rebooting
controller nodes without losing galera db quorum.

The input model was initially created as a copy of
std-3cp from which the deployer node was removed and
LCM services added to the first contoller instead.

Change-Id: I83c2f8575b570898b3a70bd52163ccfe0d7afba0