Swap roles of hed3 and 4 in additional ctrl nodes (bsc#1090548)
For the dac-3cp, dac-min, and the standard input models. The
controllers in these models also use hed3 as the backup nic
for bond0 and hed4 as the public network gateway. On compute nodes,
hed3 is the public network gateway.

So the fix of Change-Id I2d43e8e7a7dce59b898649a90d32b5ed75b0d912
is done for these 3 models as well.

Change-Id: If1779b835a2a196eb23d289fc5aa1eca3beb7a62