Add glance-api rate limit to address CVE-2016-8611 (bsc#1005886)
This patch adds a rate limiter to the glance-api haproxy endpoints.
This only effects POST requests and will not impact instance launch.
It limits how many images can be created in a 60 minute window.
The default value is 600 connections per 60 minute window which
should cover most normal glance-api use cases. Once the rate has been
exceeded, the user is locked out for the duration of the 60 minute
interval. This value can be overriden in the the input model by editing
the control_plane.yml file as follows:

             - glance-api:
                ha_mode: false
                glance_stores: 'file'
                glance_default_store: 'file'
                glance_rate_limit: <limit>

Change-Id: Ib82b61c95b228f0655365d9d3f168c70d8d8ae00