Remove manila-share dependency to keystone-client (bsc#1109264)
The manila-share service component which is associated with compute
nodes incorrectly requires the keystone-client service component,
which is causing a failure on RHEL compute nodes because zypper
is incorrectly used to install the client package.

NOTE: while this doesn't fix the root cause, it does correct the
condition that introduced the failure.

Change-Id: I964da698b910925232c534dedf1a5ce63daa89c1