SCRD-7639 Remove glance-registry from input models (SCRD-7639)
While we no longer support deploying the glance-register in the
glance-ansible code base we haven't remove the support from the
ardana-input-model repo for this service component.

This patch removes all references to this service compnent from
the service component lists in all input models, both examples
and CI, and updates the definition of the glance-registry service
component definition to generate a warning that the customer can
clear by removing any glance-registry service component entries
in their input model.

Depends-On: I1643fa70c2f7fdc0adf931b227593a23b144e9dd

Change-Id: I6faf55e4558cabc9e05301058a84cb5f2e14cdc5