Fix fail-over/-back behavior of haproxy for galera (bsc#1122875)
With the current setup it could happen, that after a database failover
connections are spread across multiple backends (which can cause DB deadlocks
with galera). This commit adjusts the backend settings for haproxy to
close all active sessions when the backend check is failing (to
fail-over everything to a backup node).

Additionally it adds a sticktable to keep the connections going to the
backup backup even after the failing backend comes back. This is to
avoid needless connection drop (and aborted transactions) after a failed
backend gets back online.

Change-Id: Iab2c5a1eed35a0935bf1381891d7c0e6d26ecc88